Add price_helper & refactor load_html_k_ruoka + update load_html_s_kaupat
